extract ContextMenuInner.svelte
